+++ src/main/java/com/example/demo/service/DaoTemplate.java
@@ -71,8 +71,14 @@ public class DaoTemplate implements CodeTemplate {
             }
 
             String fileName = rootPath + parentPath +"/"+ module + "/" + table.getClassName() + "Mapper.java";
-            if(codeConfigConstant.getWrite())
-            FileUtils.write(new File(fileName),result,"utf-8");
+            File file = new File(fileName);
+            // 不覆盖
+            if(file.exists() && codeConfigConstant.getWrite()){
+                FileUtils.write(new File(fileName),result,"utf-8");
+            }
+            if(!file.exists()){
+                FileUtils.write(new File(fileName),result,"utf-8");
+            }
         } catch (IOException e) {
             e.printStackTrace();
         } catch (TemplateException e) {
