+++ src/main/java/com/example/demo/service/EntityDeleteTemplate.java
@@ -68,8 +68,14 @@ public class EntityDeleteTemplate implements CodeTemplate {
             }
 
             String fileName = rootPath+parentPath +"/"+ module + "/" + table.getClassName() + "DeleteParam.java";
-            if(codeConfigConstant.getWrite())
-            FileUtils.write(new File(fileName),result,"utf-8");
+            File file = new File(fileName);
+            // 不覆盖
+            if(file.exists() && codeConfigConstant.getWrite()){
+                FileUtils.write(new File(fileName),result,"utf-8");
+            }
+            if(!file.exists()){
+                FileUtils.write(new File(fileName),result,"utf-8");
+            }
         } catch (IOException e) {
             e.printStackTrace();
         } catch (TemplateException e) {
