+++ src/main/java/com/example/demo/service/EntityTemplate.java
@@ -69,8 +69,14 @@ public class EntityTemplate implements CodeTemplate {
 
             String fileName = rootPath+parentPath +"/"+ module + "/" + table.getClassName() + ".java";
 
-            if(codeConfigConstant.getWrite())
-            FileUtils.write(new File(fileName),result,"utf-8");
+            File file = new File(fileName);
+            // 不覆盖
+            if(file.exists() && codeConfigConstant.getWrite()){
+                FileUtils.write(new File(fileName),result,"utf-8");
+            }
+            if(!file.exists()){
+                FileUtils.write(new File(fileName),result,"utf-8");
+            }
         } catch (IOException e) {
             e.printStackTrace();
         } catch (TemplateException e) {
